New package: kvantum-1.1.5
The package at void-packages seems to no longer ship the Qt5 support, so I'm packaging it here with it.
It should be identical to the one packaged on Void. The difference is the subpackage.
I wanted to just package the Qt5 package but it would require some patching. It was easier to package both Qt5/6.